% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aic.R
\name{aic}
\alias{aic}
\title{Akaike information criterion}
\usage{
aic(m, lambda, delta, alphaLoss = -1, filt = c("cutoff"))
}
\arguments{
\item{m}{Integer stopping index.}

\item{lambda}{Vector of decreasing, strictly positive entries of the
diagonal design matrix.}

\item{delta}{Numeric noise level.}

\item{filt}{Character string giving the filter to be used.}

\item{alpha}{Numeric smoothing parameter.}
}
\value{
Returns the value of the squared bias at index \code{m}.
}
\description{
Computes the AIC of the \code{m}-th filter estimator.
}
