% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{claOracle}
\alias{claOracle}
\title{Classical oracle}
\usage{
claOracle(lambda, mu, delta, alpha = -1, filt = "cutoff")
}
\arguments{
\item{lambda}{Vector of decreasing, strictly positive entries of the
diagonal design matrix.}

\item{mu}{Vector valued input signal.}

\item{delta}{Numeric noise level.}

\item{alpha}{Numeric smoothing parameter. \code{alpha} = -1 gives
the strong classical oracle. \code{alpha} = 0 gives the weak classical
oracle.}

\item{filt}{Character string designating the filter to be used. filt should
be one of "cutoff" or "landw".}
}
\value{
Returns the classical oracle index as an integer.
}
\description{
Computes the classical oracle index.
}
