% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fEst.R
\name{fEst}
\alias{fEst}
\title{Filter estimator}
\usage{
fEst(m, Y, lambda, filt = c("cutoff"))
}
\arguments{
\item{m}{Integer stopping index.}

\item{Y}{Numeric vector of observed data.}

\item{lambda}{Numeric vector of decreasing, strictly positive entries of
the diagonal design matrix.}

\item{filt}{Character string giving the filter to be used.}
}
\value{
Returns a numeric vector estimating the underlying signal.
}
\description{
Computes the filter estimator with stopping index \code{m} for given data
\code{Y} and diagonal design matrix \code{lambda}.
}
