% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{landw}
\alias{landw}
\title{Landweber iteration}
\usage{
landw(m, Y, lambda)
}
\arguments{
\item{m}{Integer stopping index.}

\item{Y}{Numeric vector of observed data.}

\item{lambda}{Numeric vector of decreasing, strictly positive entries of
the diagonal design matrix.}
}
\value{
Returns the m-th iterate of the Landweber iteration.
}
\description{
Computes the Landweber iteration for a diagon
}
