% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sTime}
\alias{sTime}
\title{Residual stopping time}
\usage{
sTime(Y, lambda, alpha, kappa, filt = "cutoff")
}
\arguments{
\item{Y}{Numeric vector of observed data.}

\item{lambda}{Numeric vector of decreasing, strictly positive entries of the
diagonal design matrix.}

\item{alpha}{Numeric smoothing index for the residuals.}

\item{kappa}{Strictly positive stopping value.}

\item{filt}{Character string giving the filter to be used.}
}
\value{
Returns the integer value of the smoothed residual based stopping
  time.
}
\description{
Computes the residual based stopping time for a given data
vector \code{Y} and a given diagonal design matrix \code{lambda}.
}
