% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance.R
\name{variance}
\alias{variance}
\title{Variance}
\usage{
variance(m, lambda, delta, alpha = -1, filt = c("cutoff", "landw"))
}
\arguments{
\item{m}{Integer stopping index.}

\item{lambda}{Vector of decreasing, strictly positive entries of the
diagonal design matrix.}

\item{delta}{Numeric noise level.}

\item{alpha}{Numeric smoothing parameter. \code{alpha} = -1 gives
the strong bias. \code{alpha} = 0 gives the weak bias.}

\item{filt}{Character string designating the filter to be used. filt should
be one of "cutoff" or "landw".}
}
\value{
Returns the value of the variancee at index \code{m}.
}
\description{
Computes the variance of the \code{m}-th filter estimator, a given diagonal
design matrix \code{lambda} and noise level delta.
}
