% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aic.R
\name{aic}
\alias{aic}
\title{Akaike information criterion}
\usage{
aic(m, Y, lambda, delta, alphaLoss = -1, filt = c("cutoff", "landw"))
}
\arguments{
\item{m}{Integer stopping index.}

\item{Y}{Numeric vector of observed data.}

\item{lambda}{Vector of decreasing, strictly positive entries of the
diagonal design matrix.}

\item{delta}{Numeric noise level.}

\item{alphaLoss}{Numeric smoothing parameter for the loss.}

\item{filt}{Character string giving the filter to be used.}
}
\value{
Returns the value of the AIC at index \code{m}.
}
\description{
Computes the AIC of the \code{m}-th filter estimator.
}
