% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{balOracle}
\alias{balOracle}
\title{Balanced oracle}
\usage{
balOracle(lambda, mu, delta, alpha = -1, filt = "cutoff")
}
\arguments{
\item{lambda}{Vector of decreasing, strictly positive entries of the
diagonal design matrix.}

\item{mu}{Vector valued input signal.}

\item{delta}{Strictly positive level of random noise.}

\item{alpha}{Numeric smoothing parameter.}

\item{filt}{Character string giving the filter to be used.}
}
\value{
Returns the strong balanced oracle time as an integer.
}
\description{
Computes the balanced oracle time.
}
