% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bias2}
\alias{bias2}
\title{Squared bias}
\usage{
bias2(lambda, mu, alpha = -1, filt = "cutoff")
}
\arguments{
\item{lambda}{Vector of decreasing, strictly positive entries of the
diagonal design matrix.}

\item{mu}{Vector valued input signal.}

\item{alpha}{Numeric smoothing parameter.}

\item{filt}{Character string giving the filter to be used.}
}
\value{
Returns all values of the squared bias in a double vector.
}
\description{
Computes the squared bias of the filter estimator a given diagonal design
matrix \code{lambda} and signal ' \code{mu}.
}
