% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bias2.R
\name{bias2}
\alias{bias2}
\title{Squared bias}
\usage{
bias2(m, lambda, mu, alpha = -1, filt = c("cutoff", "landw"))
}
\arguments{
\item{m}{Integer stopping index.}

\item{lambda}{Vector of decreasing, strictly positive entries of the
diagonal design matrix.}

\item{mu}{Vector valued input signal.}

\item{alpha}{Numeric smoothing parameter.}

\item{filt}{Character string giving the filter to be used.}
}
\value{
Returns the value of the squared bias at index \code{m}.
}
\description{
Computes the squared bias of the \code{m}-th filter estimator, a given
diagonal design matrix \code{lambda} and signal \code{mu}.
}
