% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{claOracle}
\alias{claOracle}
\title{Classical oracle}
\usage{
claOracle(lambda, mu, delta, alpha = -1, filt = "cutoff")
}
\arguments{
\item{lambda}{Vector of decreasing, strictly positive entries of the
diagonal design matrix.}

\item{mu}{Vector valued input signal.}

\item{delta}{Strictly positive level of random noise.}

\item{alpha}{Numeric smoothing parameter.}

\item{filt}{Character string giving the filter to be used. This must match
one of "cutoff".}
}
\value{
Returns the classical oracle time as an integer.
}
\description{
Computes the classical oracle time.
}
