% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance.R
\name{variance}
\alias{variance}
\title{Variance}
\usage{
variance(m, lambda, delta, alpha = -1, filt = c("cutoff"))
}
\arguments{
\item{m}{Integer stopping index.}

\item{lambda}{Vector of decreasing, strictly positive entries of the
diagonal design matrix.}

\item{delta}{Numeric noise level.}

\item{alpha}{Numeric smoothing parameter.}

\item{filt}{Character string giving the filter to be used.}
}
\value{
Returns all values of the variancee in a double vector.
}
\description{
Computes the variance of the filter estimator a given diagonal design
matrix \code{lambda} and noise level delta.
}
